requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
boto3>=1.34.0
botocore>=1.34.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
